# Move proto wire seeking logic to util package