--- a/server/util/bazel_request/BUILD
+++ b/server/util/bazel_request/BUILD
@@ -7,9 +7,9 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//proto:remote_execution_go_proto",
+        "//server/util/pbwireutil",
         "//server/util/status",
         "@org_golang_google_grpc//metadata",
-        "@org_golang_google_protobuf//encoding/protowire",
         "@org_golang_google_protobuf//proto",
     ],
 )