I found myself about to copy this logic again (for a WIP PR), but since we have this loop in a few places now, and it's somewhat error prone (e.g. remembering to check `n < 0`), I figured it's worth a util package.

**Related issues**: N/A