@bduffany 2023-07-14
> Since we extract the method into a separate util package, how about adding some unit tests?

Done!