@bduffany 2023-07-14
> Do you want to run lulu's pebble_cache performance tests before and after this change to make sure performance has not changed?

Before:

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7950X 16-Core Processor            
BenchmarkGetMulti-32                 634           1721611 ns/op          526312 B/op       7490 allocs/op
BenchmarkFindMissing-32             2712            401440 ns/op          207550 B/op       4926 allocs/op
BenchmarkContains1-32             188442              6580 ns/op            2507 B/op         59 allocs/op
BenchmarkSet-32                    17780             60347 ns/op            9515 B/op        133 allocs/op
```

After:

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7950X 16-Core Processor            
BenchmarkGetMulti-32                 703           1680752 ns/op          525717 B/op       7483 allocs/op
BenchmarkFindMissing-32             2775            404533 ns/op          207523 B/op       4925 allocs/op
BenchmarkContains1-32             182521              6471 ns/op            2507 B/op         59 allocs/op
BenchmarkSet-32                    18682             58683 ns/op            9467 B/op        133 allocs/op
```

I ran the benchmarks a few times and there appears to be a lot of variability in these numbers, so I think the differences aren't significant.